feat: combine SidebarIcons files
